django/requirements.txt
Cython==0.20.1
Django==1.6
MySQL-python==1.2.5
Pillow==2.3.0
RandomWords==0.1.12
South==0.8.4
django-cors-headers==0.12
django-filter==0.7
django-localflavor==1.0
django-nose==1.2
django-oauth2-provider==0.2.6
django-rest-swagger==0.1.13
django-websocket-redis==0.2.3
djangorestframework==2.3.12
factory-boy==2.3.1
gevent==1.0
greenlet==0.4.2
nodeenv==0.7.2
nose==1.3.0
oauth==1.0.1
pytz==2014.1
redis==2.9.1
requests==2.2.1
shortuuid==0.4
uWSGI==2.0.1
ujson==1.33
websocket-client==0.12.0
wsgiref==0.1.2
